Rename deprecated Chat class to ChatHandle, drop ChatHandle class
