+++ src/Customers/Entry/ChatHandle.php
@@ -4,6 +4,89 @@ declare(strict_types=1);
 
 namespace HelpScout\Api\Customers\Entry;
 
-class ChatHandle extends Chat
+use HelpScout\Api\Assert\Assert;
+use HelpScout\Api\Entity\Extractable;
+use HelpScout\Api\Entity\Hydratable;
+
+class ChatHandle implements Extractable, Hydratable
 {
+    /**
+     * @var int|null
+     */
+    private $id;
+
+    /**
+     * @var string|null
+     */
+    private $value;
+
+    /**
+     * @var string|null
+     */
+    private $type;
+
+    public function hydrate(array $data, array $embedded = [])
+    {
+        if (isset($data['id'])) {
+            $this->setId($data['id']);
+        }
+
+        $this->setValue($data['value'] ?? null);
+        $this->setType($data['type'] ?? null);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function extract(): array
+    {
+        return [
+            'value' => $this->getValue(),
+            'type' => $this->getType(),
+        ];
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function setId(int $id): ChatHandle
+    {
+        Assert::greaterThan($id, 0);
+
+        $this->id = $id;
+
+        return $this;
+    }
+
+    public function getValue(): ?string
+    {
+        return $this->value;
+    }
+
+    /**
+     * @param string|null $value
+     */
+    public function setValue($value): ChatHandle
+    {
+        $this->value = $value;
+
+        return $this;
+    }
+
+    public function getType(): ?string
+    {
+        return $this->type;
+    }
+
+    /**
+     * @param string|null $type
+     */
+    public function setType($type): ChatHandle
+    {
+        $this->type = $type;
+
+        return $this;
+    }
 }
